chore: delete dead internal/cache; unify SQLite time parsing in internal/db
internal/cache had zero imports since the SQLite migration. The identical
sqliteTimeFormats list + parse loop lived in both repo/sqlite.go and
sync/worker.go; db.ParseSQLiteTime/ParseSQLiteTimeAny now own every format a
timestamp can arrive in.
